% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsLCA_count.R
\name{ngsLCA_count}
\alias{ngsLCA_count}
\title{Count Reads and Taxa Numbers}
\usage{
ngsLCA_count(path, run = "run01")
}
\arguments{
\item{path}{working directory, same to \code{\link{ngsLCA_profile}}.}

\item{run}{name of the run, default is "run01".}
}
\value{
Text files and figures showing reads number and taxa number.
}
\description{
Count the reads number and taxa number after each filtering and de-contaminating, as well as in each taxa group. Results will be generated into "path/run/counts/"
}
\examples{
ngsLCA_count(path=system.file("extdata","lca_files",package="ngsLCA"),
             run="run01")


## This will generate tables and figures showing
## the numbers of reads and taxa for "run01". Results
## will be in "path/run01/counts/".


}
