% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsLCA_group.R
\name{ngsLCA_group}
\alias{ngsLCA_group}
\title{Group taxa in the combined taxa profile}
\usage{
ngsLCA_group(
  path,
  run = "run01",
  group.name = "Viruses,Archaea,Bacteria,Fungi,Viridiplantae,Metazoa",
  threshold.perGroup = 0
)
}
\arguments{
\item{path}{working directory, same to \code{\link{ngsLCA_profile}}.}

\item{run}{name of the run, default is "run01".}

\item{group.name}{a comma seprated vector listing the taxonomic units that will be used for grouping taxa, need to be as the scientific names of NCBI taxonomy (https://www.ncbi.nlm.nih.gov/taxonomy); default is "Viruses,Archaea,Bacteria,Fungi,Viridiplantae,Metazoa".}

\item{threshold.perGroup}{minimum reads percentage (to the total reads number of each group) required for confirming a taxon in each group of each sample, ranging from 0 to 1.}
}
\description{
Group the combined taxa profile into user-defined taxa units. Results will be in 'pth/run/taxonomic_profiles/taxa_groups/'.
}
\examples{
ngsLCA_group(path=system.file("extdata","lca_files",package="ngsLCA"),
             run="run01",
             group.name="Viridiplantae,Metazoa",
             threshold.perGroup=0.01)

## This will extract all taxa under Viridiplantae and Metazoa
## from the combined taxa profile of 'run01', and filter
## these 2 taxa profiles by threshold.perGroup to generate 2
## files named "Viridiplantae.txt" and "Metazoa.txt" in
## "extdata/lca_files/run01/taxonomic_profiles/taxa_groups/".


}
