% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsLCA_meganFile.R
\name{ngsLCA_meganFile}
\alias{ngsLCA_meganFile}
\title{Generate Megan files}
\usage{
ngsLCA_meganFile(path, run = "run01")
}
\arguments{
\item{path}{working directory, same to \code{\link{ngsLCA_profile}}.}

\item{run}{name of the run, default is "run01".}
}
\description{
Generate files that can be inputted into Megan for taxa profiling. Results will be in 'path/run/megan/'
}
\examples{
ngsLCA_meganFile(path=system.file("extdata","lca_files",package="ngsLCA"),
                 run="run01")


## This will generate Megan format taxa profiles for
## the complete taxa profile as well as taxa groups
## and ranks (if available) of 'run01'. Resutls will
## be in extdata/lca_files/run01/megan/'. These files
## can be directly used as input csv files in Megan
## for taxa profiling.


}
