% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsLCA_profile.R
\name{ngsLCA_profile}
\alias{ngsLCA_profile}
\title{Generate a combined taxa profile}
\usage{
ngsLCA_profile(path, run = "run01", remove.sample = NULL, metadata = NULL)
}
\arguments{
\item{path}{working directory, which should contain lca files that will be processed.}

\item{run}{run name, a folder that will be created for storing generated files and results; default is "run01".}

\item{remove.sample}{a comma separated vector listing file names indicating samples that will NOT be included for analysis, optional.}

\item{metadata}{path to sample metadata, an example at github.com/miwipe/ngsLCA/blob/master/R_script/metadata.txt, optional.}
}
\description{
Read lca files in the appointed directory to generate a combined taxa profile named 'complete_profile.txt' in folder 'path/run/taxonomic_profiles/'.
}
\examples{
ngsLCA_profile(path="working_directory/",
               run="run01",
               remove.sample="sample1.lca,sample2",
               metadata="path_to_metadata/metadata.txt")

## This will combine all lca files in "working_directory/"
## to generate a file named 'complete_profile.txt' in
## 'working_directory/run01/taxonomic_profiles/', with
## replacing the file names by sample names supplied in
## metadata.txt, and removing sample1 and sample2.


}
