% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsLCA_stratplot.R
\name{ngsLCA_stratplot}
\alias{ngsLCA_stratplot}
\title{Generate Taxa Abundance Stratplots}
\usage{
ngsLCA_stratplot(path, run = "run01", taxa.number = 10)
}
\arguments{
\item{path}{working directory, same to \code{\link{ngsLCA_profile}}.}

\item{run}{name of the run, default is "run01".}

\item{taxa.number}{maximum number of taxa will be shown in stratplots; default is 10.}
}
\value{
Taxa abundance stratplots.
}
\description{
Generate stratplots for taxa profiles into "path/run/stratplot/" to show the reads abundance.
}
\examples{
ngsLCA_stratplot(path=system.file("extdata","lca_files",package="ngsLCA"),
                 run="run01",
                 taxa.number=10)


## This will generate stratplots for the complete taxa profile
## as well as taxa groups and ranks (if available) of "run01",
## with showing the 10 most abundant taxa in each profile.
## Results will be in "path/run01/stratplot/".


}
