% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsLCA_deContam.R
\name{ngsLCA_deContam}
\alias{ngsLCA_deContam}
\title{Remove contamination taxa}
\usage{
ngsLCA_deContam(
  path,
  run = "run01",
  control_path,
  threshold.1_control = 2,
  threshold.2_control = 0,
  threshold.3_control = 5
)
}
\arguments{
\item{path}{working directory, same to 'ngsLCA_profile'.}

\item{run}{name of the run, default is "run01".}

\item{control_path}{path to the directory that contains all blank control lca files.}

\item{threshold.1_control}{for generating contamination list, minimum reads number required for keeping a taxon in each blank control; default is 2.}

\item{threshold.2_control}{for generating contamination list, minimum read percentage (to the total reads number of a blank control) required for keeping a taxon in each blank control, ranging from 0 to 1; default is 0.}

\item{threshold.3_control}{for generating contamination list, minimum sum of reads across all blank controls required for keeping a taxon; default is 5.}
}
\value{

}
\description{
Generate a contamination list from blank control lca files, and subtract listed taxa from the combined taxa profile. Running this function will update the existing 'complete_profile.txt'.
}
\examples{
ngsLCA_deContam(path="working_directory/",
                run="run01",
                control_path="working_directory/blank_control_folder/",
                threshold.1_control=2,
                threshold.2_control=0,
                threshold.3_control=5)

## This will combine the blank control lca files within
## folder "working_directory/blank_control_folder/"
## to generate a contamination list, and then filter
## this list using the 3 control thresholds supplied.
## The listed taxa will be subtracted from the combined
## taxa profile.


}
