% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsLCA_heatmap.R
\name{ngsLCA_heatmap}
\alias{ngsLCA_heatmap}
\title{Generate taxa abundance heatmaps}
\usage{
ngsLCA_heatmap(path, run = "run01", taxa.number = 30)
}
\arguments{
\item{path}{working directory, same to 'ngsLCA_profile'.}

\item{run}{name of the run, default is "run01".}

\item{taxa.number}{maximum number of taxa will be shown in heatmaps; default is 30.}
}
\description{
Generate heatmaps for taxa profiles into 'path/run/heatmap/' to show the reads abundance of each taxon.
}
\examples{
ngsLCA_heatmap(path="working_directory/",
               run="run01",
               taxa.number=20)


## This will generate heatmaps for the complete taxa profile
## as well as taxa groups and ranks (if available) of 'run01'
## in 'working_directory/run01/heatmap/', with showing the 20
## most abundant taxa in each profile.


}
