% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsLCA_NMDS.R
\name{ngsLCA_NMDS}
\alias{ngsLCA_NMDS}
\title{Taxa Profiles NMDS}
\usage{
ngsLCA_NMDS(path, run = "run01", dimension = 3, trymax = 100)
}
\arguments{
\item{path}{working directory, same to \code{\link{ngsLCA_profile}}.}

\item{run}{name of the run, default is "run01".}

\item{dimension}{dimension of the NMDS; default is 3.}

\item{trymax}{maximum number of random starts in search of convergent solutions for NMDS; default is 100.}
}
\value{
R data files for NMDS results, figures for NMDS stressplots, cluster figures for the first 2 NMDS components.
}
\description{
Perform Non-metric Multidimensional Scaling (NMDS) on taxa profiles. Results will be in "path/run/NMDS/". See vegan::metaMDS for details.
}
\examples{
\dontrun{ngsLCA_NMDS(path=system.file("extdata","lca_files",package="ngsLCA"),
            run="run01",
            dimension=3,
            trymax=1000)}


## This will perform NMDS on the complete taxa
## profile as well as taxa groups and ranks
## (if available) of "run01". Results will be in
## "path/run01/NMDS/".


}
