% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsLCA_rarefy.R
\name{ngsLCA_rarefy}
\alias{ngsLCA_rarefy}
\title{Rarefaction taxa reads abundance}
\usage{
ngsLCA_rarefy(path, run = "run01")
}
\arguments{
\item{path}{working directory, same to 'ngsLCA_profile'.}

\item{run}{name of the run, default is "run01".}
}
\value{

}
\description{
Generate rarefaction curves for taxa profiles in 'path/run/rarefaction/'. See vegan::rarefy for details.
}
\examples{
ngsLCA_rarefy(path="working_directory/",
              run="run01")

## This will generate rarefaction curves for the
## complete taxa profile as well as taxa groups
## and ranks (if available) of 'run01'. Results
## will be in 'working_directory/run01/rarefaction/'.


}
